+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ##### Bug Fixes
 
 - Fix unused import false-positive for when a declaration's associated types are declared in different module than module that declares the declaration.
+- Fix unused import false-positive for modules that export unindexed modules.
 
 ## 3.2.0 (2025-06-27)